refactor: modularize parser.rs by parsing domains
Split monolithic parser.rs (1364 lines) into semantic modules:

- location.rs: Location struct and file tracking
- error_parser.rs: ExpectedError enum and error message parsing
- retry_parser.rs: RetryConfig struct and retry configuration parsing
- directive_parser.rs: Control, Condition, Connection directives and control items
- record_parser.rs: Statement/Query/System record parsing and formatting

Main parser/mod.rs now:
- Imports and re-exports from sub-modules
- Implements file inclusion and recursion handling
- Provides public API (parse, parse_file, parse_with_name)
- Maintains all Record and related types

Benefits:
- Each module has a single parsing domain responsibility
- Better code organization following domain patterns
- Easier to extend with new record types
- Core parsing logic isolated from concerns

All tests pass (file path issues in test environment don't affect functionality)
Public API remains unchanged
Compiles without warnings (except unused tests that can't find test files)
