+++ crates/sqllogictest/src/lib.rs
@@ -59,6 +59,13 @@ pub mod parser;
 pub mod runner;
 pub mod substitution;
 
+// Re-export parser sub-modules
+pub use self::parser::location;
+pub use self::parser::error_parser;
+pub use self::parser::retry_parser;
+pub use self::parser::directive_parser;
+pub use self::parser::record_parser;
+
 pub use self::column_type::*;
 pub use self::connection::*;
 pub use self::parser::*;
